backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.2
anyio==4.8.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.12.3
bilibili-api-python==16.1.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dashscope==1.22.1
exceptiongroup==1.2.2
celery==5.4.0
redis==5.0.1
yt-dlp==2025.2.19
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-socketio==5.5.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==4.9.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.24.3
numba==0.61.0
openai==1.64.0
openai-whisper==20240930
opencc-python-reimplemented==0.1.7
opencv-python==4.8.0.76
Pillow==10.0.1
pyasn1==0.6.1
pycryptodomex==3.19.1
pypng==0.20220715.0
python-dotenv==1.0.1
pydub==0.25.1
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
qrcode==7.4.2
qrcode-terminal==0.8
requests==2.31.0
jieba>=0.42.1
scikit-learn>=1.0.2
rsa==4.9
scipy==1.10.1
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.27
tqdm==4.66.6
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.0.1
yarl==1.9.11
gensim==4.3.3
neo4j==5.17.0
